# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Local PII Redaction API
  version: 0.1.0
  description: Local-first session job and artifact control plane. Artifact bytes are never embedded in JSON.
x-implementation-status: session-artifact-scan-redact-and-verified-download
servers:
  - url: http://127.0.0.1:{port}
    description: Local loopback profile
    variables:
      port:
        default: "7314"
paths:
  /health/live:
    get:
      operationId: getLiveness
      security: []
      responses:
        "204": { description: Process is alive }
  /health/ready:
    get:
      operationId: getReadiness
      security: []
      responses:
        "204": { description: Required local dependencies and contracts are compatible }
        "503":
          description: A required dependency is unavailable
          content:
            application/json:
              schema:
                $ref: "./schemas/common/errors.schema.json"
  /v1/capabilities:
    get:
      operationId: getCapabilities
      responses:
        "200":
          description: Versioned local capability manifest
          content:
            application/json:
              schema:
                $ref: "./schemas/capabilities/capability-manifest.schema.json"
  /v1/policies:
    get:
      operationId: getPolicyCatalog
      responses:
        "200":
          description: Pinned privacy-safe policy references available to the local application
          content:
            application/json:
              schema:
                $ref: "./schemas/policy/policy-catalog.schema.json"
  /v1/preview/scan:
    post:
      operationId: scanEphemeralPreview
      description: Authenticated process-local rules scan. Request bytes are not added to an artifact or job store.
      parameters:
        - name: format
          in: query
          required: true
          schema: { enum: [text, markdown] }
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, contentMediaType: application/octet-stream }
      responses:
        "200":
          description: Privacy-minimized aggregate preview result
          content:
            application/json:
              schema:
                $ref: "./schemas/jobs/preview-scan-report.schema.json"
  /v1/preview/review:
    post:
      operationId: reviewEphemeralPreview
      description: Authenticated process-local rules scan returning bounded value-free detection and conflict locations.
      parameters:
        - name: format
          in: query
          required: true
          schema: { enum: [text, markdown] }
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, contentMediaType: application/octet-stream }
      responses:
        "200":
          description: Bounded value-free detection and conflict review result
          content:
            application/json:
              schema:
                $ref: "./schemas/jobs/preview-review-report-v2.schema.json"
  /v1/artifacts:
    post:
      operationId: createSessionArtifact
      description: Creates bounded process-local input-artifact metadata. No filename or storage locator is accepted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/artifacts/create-artifact-request.schema.json"
      responses:
        "201":
          description: A staged session artifact was created
          content:
            application/json:
              schema:
                $ref: "./schemas/artifacts/artifact.schema.json"
  /v1/artifacts/{artifactId}/content:
    parameters:
      - name: artifactId
        in: path
        required: true
        schema: { type: string, pattern: '^art_[0-9A-HJKMNP-TV-Z]{26}$' }
    put:
      operationId: uploadSessionArtifactContent
      description: Supplies bounded bytes once and verifies them against the declared length and digest.
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, contentMediaType: application/octet-stream, maxLength: 8388608 }
      responses:
        "200":
          description: Artifact bytes were accepted as immutable for this process-local session
          content:
            application/json:
              schema:
                $ref: "./schemas/artifacts/artifact.schema.json"
    get:
      operationId: downloadVerifiedSessionArtifact
      description: Downloads a server-authorized, verified process-local redacted output. Input artifacts are never downloadable through this route.
      responses:
        "200":
          description: Exact verified redacted bytes with a generic attachment name
          content:
            text/plain:
              schema: { type: string, contentMediaType: text/plain, maxLength: 16777216 }
            text/markdown:
              schema: { type: string, contentMediaType: text/markdown, maxLength: 16777216 }
        "404":
          description: Artifact is unavailable in the caller's authorization view
          content:
            application/json:
              schema:
                $ref: "./schemas/common/errors.schema.json"
  /v1/jobs/{jobId}:
    parameters:
      - name: jobId
        in: path
        required: true
        schema: { type: string, pattern: '^job_[0-9A-HJKMNP-TV-Z]{26}$' }
    get:
      operationId: getJob
      responses:
        "200":
          description: Current job state
          content:
            application/json:
              schema:
                $ref: "./schemas/jobs/job.schema.json"
        "404":
          description: Job is unavailable in the caller's authorization view
          content:
            application/json:
              schema:
                $ref: "./schemas/common/errors.schema.json"
  /v1/jobs:
    post:
      operationId: createJob
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
            pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "./schemas/jobs/create-job-request.schema.json"
                - $ref: "./schemas/jobs/create-job-request-v2.schema.json"
                - $ref: "./schemas/jobs/create-job-request-v3.schema.json"
      responses:
        "201":
          description: A new metadata-only, process-local scan, or process-local redaction job was created
          content:
            application/json:
              schema:
                $ref: "./schemas/jobs/job.schema.json"
        "200":
          description: The original creation result was replayed for the idempotency key
          content:
            application/json:
              schema:
                $ref: "./schemas/jobs/job.schema.json"
        "409":
          description: The idempotency key conflicts with a different request
          content:
            application/json:
              schema:
                $ref: "./schemas/common/errors.schema.json"
  /v1/jobs/{jobId}/events:
    parameters:
      - name: jobId
        in: path
        required: true
        schema: { type: string, pattern: '^job_[0-9A-HJKMNP-TV-Z]{26}$' }
    get:
      operationId: listJobEvents
      parameters:
        - name: after
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
      responses:
        "200":
          description: Bounded cursor page of privacy-minimized job events
          content:
            application/json:
              schema:
                $ref: "./schemas/jobs/job-event-page.schema.json"
        "404":
          description: Job is unavailable in the caller's authorization view
          content:
            application/json:
              schema:
                $ref: "./schemas/common/errors.schema.json"
  /v1/jobs/{jobId}/detections:
    parameters:
      - name: jobId
        in: path
        required: true
        schema: { type: string, pattern: '^job_[0-9A-HJKMNP-TV-Z]{26}$' }
    get:
      operationId: listJobDetections
      parameters:
        - name: cursor
          in: query
          schema: { type: integer, minimum: 0, maximum: 10000, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
      responses:
        "200":
          description: Bounded value-free resolved detection and conflict page for a completed scan
          content:
            application/json:
              schema:
                $ref: "./schemas/jobs/detection-page.schema.json"
  /v1/jobs/{jobId}/output:
    parameters:
      - name: jobId
        in: path
        required: true
        schema: { type: string, pattern: '^job_[0-9A-HJKMNP-TV-Z]{26}$' }
    get:
      operationId: getVerifiedJobOutput
      description: Returns minimized metadata only after a redaction job reaches VERIFIED.
      responses:
        "200":
          description: Server-authorized verified process-local output metadata
          content:
            application/json:
              schema:
                $ref: "./schemas/artifacts/artifact.schema.json"
        "404":
          description: Job or output is unavailable in the caller's authorization view
          content:
            application/json:
              schema:
                $ref: "./schemas/common/errors.schema.json"
  /v1/jobs/{jobId}/cancellation:
    parameters:
      - name: jobId
        in: path
        required: true
        schema: { type: string, pattern: '^job_[0-9A-HJKMNP-TV-Z]{26}$' }
    post:
      operationId: cancelJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/jobs/cancel-job-request.schema.json"
      responses:
        "200":
          description: Cancellation was accepted with a new aggregate revision
          content:
            application/json:
              schema:
                $ref: "./schemas/jobs/job.schema.json"
        "409":
          description: Expected revision is stale or cancellation is not allowed in the current state
          content:
            application/json:
              schema:
                $ref: "./schemas/common/errors.schema.json"
components:
  securitySchemes:
    LocalSession:
      type: http
      scheme: bearer
      bearerFormat: opaque-local-session
security:
  - LocalSession: []